move ChangeFactory
