+++ src/main/java/spoon/experimental/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * this is unstable code, subject to potential complete redesign, not only in snapshots but also in releases. you can use this code but you'll probably be impacted upon future changes.
+ */
+package spoon.experimental;
